% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLRShiny.R
\name{BLRShiny}
\alias{BLRShiny}
\title{Launch 'BLRShiny' Interface}
\usage{
BLRShiny()
}
\value{
Nothing
}
\description{
BLRShiny() loads interactive user interface built using R 'shiny'.
}
\details{
Start BLRShiny

The interactive user interface is to provide an easy way for binary logistic regression analysis and downloading relevant plot.
}
\examples{
if(interactive()){
library(rmarkdown)
BLRShiny()
}
}
\keyword{BLRShiny}
